refactor: getConfigFromCli
